Move ret_* functions to ret_dynamic. Surprisingly, it all worked.
