+++ Project Files/To Do.txt
@@ -17,11 +17,10 @@ VISUAL:
 DEEPER: 
 -(3,2) Sync functions are used a few places for editing. Determine if that's necessary and get rid of them.
 -(3,2) double check that all forms are checked on both sides
--(2,2) Move ret_ functions to their own file (Then res would need to be passed too)
 -(2,3) ?Maybe use ajax instead of form for linkloading-trainer?
 -(2,4) ?use json? ajaj? instead of ajax and .data and xml?
 -(2,5) ?Convert everything in the session to 307 redirects?
--(1,2) ?move reports out of error and into report folder?
+-(1,2) ?move reports out of error log and into report folder?
 -(1,3) ?move more things to aux?
 -(1,3) ?Should I make it check password each step of the session for safety -- would that be taxing?
 -(1,5) Switch over to a better data storage system like MySQL. (If traffic gets crazy)
